+++ packages/tests/__utils__/index.ts
@@ -3,6 +3,7 @@ export * from './faketimers.js';
 export * from './h.js';
 export * from './lifecycle.js';
 export * from './mockImageLoad.js';
+export * from './mockVideoLoad.js';
 export * from './mockIntersectionObserver.js';
 export * from './resizeWindow.js';
 export * from './wait.js';