+++ packages/tests/__utils__/mockImageLoad.ts
@@ -1,19 +1,31 @@
-let tmp;
+let descriptor;
+
+function mockImage(event) {
+  descriptor = Object.getOwnPropertyDescriptor(HTMLImageElement.prototype, 'src');
+  Object.defineProperty(HTMLImageElement.prototype, 'src', {
+    configurable: true,
+    get() {
+      return descriptor.get.call(this);
+    },
+    set(src) {
+      descriptor.set.call(this, src);
+      this.dispatchEvent(new Event(event));
+    },
+  });
+}
 
 export function mockImageLoad() {
-  tmp = global.Image;
-  global.Image = class extends tmp {
-    set src(src) {
-      this.setAttribute('src', src);
-      this.dispatchEvent(new Event('load'));
-    }
-  }
+  mockImage('load');
+}
+
+export function mockImageLoadError() {
+  mockImage('error');
 }
 
 export function unmockImageLoad() {
-  if (tmp) {
-    global.Image = tmp;
+  if (descriptor) {
+    Object.defineProperty(HTMLImageElement.prototype, 'src', descriptor);
   }
 
-  tmp = null;
+  descriptor = null;
 }